Modernise .cabal file
The 0.6.8.1 release was rather pointless, as it both redundant and
didn't have any benefit, as it didn't perform any semantic changes
that wouldn't also be possible by .cabal revisions.

Moroever, it missed the opportunity to clean up the
`.cabal` file and fix a few minor issues. This commit tries to rectify
this, so that the next proper release can contain these cleanups.

NB: There is no need to release this `.cabal` change to Hackage until
    there's other relevant changes that force a release.